+++ ll/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Implement `fmt::Display` for `SocketInterrupt`.
+
 ## [0.5.1] - 2021-02-06
 ### Fixed
 - Fixed documentation builds failing without the `embedded-hal` feature.